[refactor] move isValidElementType to react-is
